requirements.txt
asgiref             3.8.1
click               8.2.1
colorama            0.4.6
Django              5.2.3
djangorestframework 3.16.0
joblib              1.5.1
nltk                3.9.1
numpy               2.3.0
pandas              2.3.0
pip                 24.3.1
python-dateutil     2.9.0.post0
pytz                2025.2
regex               2024.11.6
six                 1.17.0
sqlparse            0.5.3
tqdm                4.67.1
tzdata              2025.2
